# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  internal_server_handlers.ClaimctlResource:
    properties:
      createdAt:
        $ref: "#/definitions/pgtype.Int8"
      id:
        type: integer
      labels:
        items: {}
        type: array
      name:
        type: string
      properties:
        $ref: "#/definitions/claimctl_com_internal_types.JSONBMap"
      spaceId:
        type: integer
      type:
        type: string
      updatedAt:
        $ref: "#/definitions/pgtype.Int8"
    type: object
...
  pgtype.Bool:
    properties:
      bool:
        type: boolean
      valid:
        type: boolean
    type: object
...
  pgtype.Int4:
    properties:
      int32:
        format: int32
        type: integer
      valid:
        type: boolean
    type: object
  pgtype.Int8:
    properties:
      int64:
        format: int64
        type: integer
      valid:
        type: boolean
    type: object
...
  pgtype.Text:
    properties:
      string:
        type: string
      valid:
        type: boolean
    type: object
...
  claimctl_com_internal_db.CreateNewResourceParams:
    properties:
      labels:
        items: {}
        type: array
      name:
        type: string
      properties:
        $ref: "#/definitions/claimctl_com_internal_types.JSONBMap"
      spaceId:
        type: integer
      type:
        type: string
    type: object
  claimctl_com_internal_db.ClaimctlResource:
    properties:
      createdAt:
        $ref: "#/definitions/pgtype.Int8"
      id:
        type: integer
      labels:
        items: {}
        type: array
      name:
        type: string
      properties:
        $ref: "#/definitions/claimctl_com_internal_types.JSONBMap"
      spaceId:
        type: integer
      type:
        type: string
      updatedAt:
        $ref: "#/definitions/pgtype.Int8"
    type: object
  claimctl_com_internal_db.ClaimctlResourceHealthConfig:
    properties:
      checkType:
        type: string
      createdAt:
        $ref: "#/definitions/pgtype.Int8"
      enabled:
        $ref: "#/definitions/pgtype.Bool"
      intervalSeconds:
        $ref: "#/definitions/pgtype.Int4"
      resourceId:
        type: integer
      retryCount:
        $ref: "#/definitions/pgtype.Int4"
      target:
        type: string
      timeoutSeconds:
        $ref: "#/definitions/pgtype.Int4"
      updatedAt:
        $ref: "#/definitions/pgtype.Int8"
    type: object
  claimctl_com_internal_db.ClaimctlResourceHealthStatus:
    properties:
      checkedAt:
        type: integer
      createdAt:
        $ref: "#/definitions/pgtype.Int8"
      errorMessage:
        $ref: "#/definitions/pgtype.Text"
      id:
        type: integer
      resourceId:
        type: integer
      responseTimeMs:
        $ref: "#/definitions/pgtype.Int4"
      status:
        type: string
    type: object
  claimctl_com_internal_db.UpdateResourceByIdParams:
    properties:
      id:
        type: integer
      labels:
        items: {}
        type: array
      name:
        type: string
      properties:
        $ref: "#/definitions/claimctl_com_internal_types.JSONBMap"
      type:
        type: string
    type: object
  claimctl_com_internal_services.ResourceWithStatus:
    properties:
      activeReservations:
        type: integer
      healthConfig:
        $ref: "#/definitions/claimctl_com_internal_db.ClaimctlResourceHealthConfig"
      healthStatus:
        $ref: "#/definitions/claimctl_com_internal_db.ClaimctlResourceHealthStatus"
      nextQueuePosition:
        type: integer
      nextUserId:
        type: integer
      queueLength:
        type: integer
      resource:
        $ref: "#/definitions/claimctl_com_internal_db.ClaimctlResource"
    type: object
  claimctl_com_internal_types.JSONBMap:
    additionalProperties:
      type: string
    type: object
host: localhost:3000
info:
  contact:
    email: support@github.com/thetaqitahmid/claimctl
    name: API Support
  description: API documentation for claimctl resource management system.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: claimctl API
  version: "1.0"
paths:
...
  /resources:
    get:
      consumes:
        - application/json
      description:
        "Retrieve a list of all resources. Note: This might return all
        resources regardless of permissions in the current implementation."
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: "#/definitions/internal_server_handlers.ClaimctlResource"
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all resources
      tags:
        - resources
    post:
      consumes:
        - application/json
      description: Create a new resource. Requires admin privileges.
      parameters:
        - description: New Resource Details
          in: body
          name: resource
          required: true
          schema:
            $ref: "#/definitions/claimctl_com_internal_db.CreateNewResourceParams"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/internal_server_handlers.ClaimctlResource"
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new resource
      tags:
        - resources
  /resources/{id}:
    delete:
      consumes:
        - application/json
      description: Delete an existing resource. Requires admin privileges.
      parameters:
        - description: Resource ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a resource
      tags:
        - resources
    get:
      consumes:
        - application/json
      description: Retrieve details of a specific resource by its ID.
      parameters:
        - description: Resource ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/internal_server_handlers.ClaimctlResource"
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a resource by ID
      tags:
        - resources
    patch:
      consumes:
        - application/json
      description: Update details of an existing resource. Requires admin privileges.
      parameters:
        - description: Resource ID
          in: path
          name: id
          required: true
          type: integer
        - description: Updated Resource Details
          in: body
          name: resource
          required: true
          schema:
            $ref: "#/definitions/claimctl_com_internal_db.UpdateResourceByIdParams"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/internal_server_handlers.ClaimctlResource"
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a resource
      tags:
        - resources
  /resources/{id}/with-status:
    get:
      consumes:
        - application/json
      description:
        Retrieve details of a specific resource including its current reservation
        status by its ID.
      parameters:
        - description: Resource ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/claimctl_com_internal_services.ResourceWithStatus"
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a resource with status by ID
      tags:
        - resources
  /resources/with-status:
    get:
      consumes:
        - application/json
      description:
        Retrieve a list of all resources including their current reservation
        status. Filtered by user permissions.
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: "#/definitions/claimctl_com_internal_services.ResourceWithStatus"
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get resources with status
      tags:
        - resources
...
swagger: "2.0"